backend/requirements.txt
fastapi[all]==0.109.2
google-genai==0.3.0
PyMuPDF==1.23.21
google-search-results==2.4.2
python-dotenv==1.0.1
arxiv==2.1.0
httpx==0.27.0
python-dateutil>=2.8.2
